Added process manager (pmgr) support
When pgbouncer is compiled with process manager support (pmgr) and
cf_pmgr_workers is larger than 1, pgbouncer will fork into manager
process and a number of worker processes. Pgbouncer the listens for
new connections and once connected it selects one of the workers
(currently round-robin) and sends the client socket to the worker
via unix domain socket the worker is listening on. The worker then
proceeds to work with the client socket as usual.

The manager process also handles signals by sending the incoming signal
to all workers.

cf_reboot, cf_daemon and cf_logfile are not currently supported when
using pmgr and cf_unix_socket_dir is required.